renamed BertLM class file
